requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-Cors==1.9.0
Flask-Login==0.2.11
Flask-OpenID==1.2.4
Flask-RESTful==0.2.12
Flask-Restless==0.13.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
PyMySQL==0.6.2
PyYAML==3.10
SQLAlchemy==0.9.4
Tempita==0.5.2
Werkzeug==0.9.6
aniso8601==0.82
argparse==1.3.0
beautifulsoup4==4.3.2
decorator==3.4.0
httplib2==0.9
itsdangerous==0.24
mimerender==0.5.4
nltk==2.0.4
numpy==1.9.0
oauth2==1.5.211
oauthlib==0.6.0
pbr==0.10.0
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
python-twitter==0.8.2
pytz==2014.2
requests==2.2.0
requests-oauthlib==0.4.0
simplejson==3.3.2
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.13
stevedore==1.1.0
tweepy==2.3.0
twitter-application-only-auth==0.3.1
virtualenv==1.10.1
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
wsgiref==0.1.2
